Move Roles & Permissions into standalone files
This should make both these enumerations more manageable
